requirements.txt
# Core dependencies for LLM inference gateway
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
transformers==4.35.2
torch==2.1.1
accelerate==0.25.0
sentencepiece==0.1.99
protobuf==4.25.1

# Observability
prometheus-client==0.19.0
structlog==23.2.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
